requirements-dev.txt
asn1crypto==0.24.0
bcrypt==3.1.6
boto3==1.14.10
cffi==1.14.1
chardet==3.0.4
Click==7.0
cryptography==3.4.6
decorator==4.0.10
docopt==0.6.1
enum34==1.1.6
fabric==2.5.0
Flask==1.1.2
flask-cors
funcsigs==1.0.2
google-api-python-client==1.9.3
gunicorn==20.0.4
httplib2==0.18.1
httpretty==0.8.10
idna==2.7
ipaddress==1.0.17
ipython==0.13.2
itsdangerous==1.1.0
Jinja2==2.10.1
livereload==1.0.0
MarkupSafe==1.1.1
mock==2.0.0
moto==0.4.25
oauth2client==4.1.3
paramiko==2.4.2
pbr==1.10.0
ply==3.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.14
pycrypto==2.6
PyNaCl==1.3.0
pyOpenSSL==20.0.1
python-slugify==1.1.2
requests==2.24.0
robotframework==3.2.1
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.4.0
rsa==3.1.4
selenium==3.141.0
sentinels==1.0.0
simplejson==3.6.5
six==1.15.0
slimit==0.8.1
tornado==3.1
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==0.4.17
uritemplate==3.0.1
urllib3==1.24.2
Werkzeug==1.0.1
xmltodict==0.10.2
gunicorn==20.0.4
